feat: add back switching to note edit page
